fetch-agent-fire-responder/paramedic/requirements.txt
uagents>=0.10.0
python-dotenv>=1.0.0
geopy>=2.4.1
supabase>=1.0.3
httpx>=0.27.0
openai>=1.30.1
groq